gateway-api: Move predicates to new predicates module
This commit moves the Predicate methods on the various reconcilers
into a new `predicates` module.

It also does some cleanup of constants and associated helpers that
are used in these functions, and moves them into the `helpers`
package.
