+++ operator/pkg/gateway-api/controller_test.go
@@ -9,297 +9,16 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cilium/hive/hivetest"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	controllerruntime "github.com/cilium/cilium/operator/pkg/controller-runtime"
-	"github.com/cilium/cilium/operator/pkg/gateway-api/helpers"
 )
-
-var controllerTestFixture = []client.Object{
-	// Cilium Gateway Class
-	&gatewayv1.GatewayClass{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "cilium",
-		},
-		Spec: gatewayv1.GatewayClassSpec{
-			ControllerName: "io.cilium/gateway-controller",
-		},
-	},
-
-	// Secret used in Gateway
-	&corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "tls-secret",
-			Namespace: "default",
-		},
-		StringData: map[string]string{
-			"tls.crt": "cert",
-			"tls.key": "key",
-		},
-		Type: corev1.SecretTypeTLS,
-	},
-
-	// Gateway with valid TLS secret
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "valid-gateway",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name:     "https",
-					Hostname: ptr.To[gatewayv1.Hostname]("example.com"),
-					Port:     443,
-					TLS: &gatewayv1.ListenerTLSConfig{
-						CertificateRefs: []gatewayv1.SecretObjectReference{
-							{
-								Name: "tls-secret",
-							},
-						},
-					},
-				},
-			},
-		},
-	},
-
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "valid-gateway-2",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name:     "https",
-					Hostname: ptr.To[gatewayv1.Hostname]("example2.com"),
-					Port:     443,
-					TLS: &gatewayv1.ListenerTLSConfig{
-						CertificateRefs: []gatewayv1.SecretObjectReference{},
-					},
-				},
-			},
-		},
-	},
-
-	// Gateway with no TLS listener
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "gateway-with-no-tls",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name: "https",
-					Port: 80,
-				},
-			},
-		},
-	},
-
-	// Gateway for TLSRoute
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "gateway-tlsroute",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name:     "tls",
-					Protocol: gatewayv1.TLSProtocolType,
-					Port:     443,
-				},
-			},
-		},
-	},
-
-	// Gateway with allowed route in same namespace only
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "gateway-from-same-namespace",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name: "https",
-					Port: 80,
-					AllowedRoutes: &gatewayv1.AllowedRoutes{
-						Namespaces: &gatewayv1.RouteNamespaces{
-							From: ptr.To(gatewayv1.NamespacesFromSame),
-						},
-					},
-				},
-			},
-		},
-	},
-
-	// Gateway with allowed routes from ALL namespace
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "gateway-from-all-namespaces",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name: "https",
-					Port: 80,
-					AllowedRoutes: &gatewayv1.AllowedRoutes{
-						Namespaces: &gatewayv1.RouteNamespaces{
-							From: ptr.To(gatewayv1.NamespacesFromAll),
-						},
-					},
-				},
-			},
-		},
-	},
-
-	// Gateway with allowed routes with selector
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "gateway-with-namespaces-selector",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name: "https",
-					Port: 80,
-					AllowedRoutes: &gatewayv1.AllowedRoutes{
-						Namespaces: &gatewayv1.RouteNamespaces{
-							From: ptr.To(gatewayv1.NamespacesFromSelector),
-							Selector: &metav1.LabelSelector{
-								MatchLabels: map[string]string{
-									"gateway": "allowed",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	},
-
-	// Gateway with allowed routes invalid namespace selector
-	&gatewayv1.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "gateway-with-invalid-namespaces-selector",
-			Namespace: "default",
-		},
-		Spec: gatewayv1.GatewaySpec{
-			GatewayClassName: "cilium",
-			Listeners: []gatewayv1.Listener{
-				{
-					Name: "https",
-					Port: 80,
-					AllowedRoutes: &gatewayv1.AllowedRoutes{
-						Namespaces: &gatewayv1.RouteNamespaces{
-							From: ptr.To(gatewayv1.NamespacesFromSelector),
-						},
-					},
-				},
-			},
-		},
-	},
-}
-
-func Test_hasMatchingController(t *testing.T) {
-	logger := hivetest.Logger(t)
-	c := fake.NewClientBuilder().WithScheme(helpers.TestScheme(helpers.AllOptionalKinds)).WithObjects(controllerTestFixture...).Build()
-	fn := hasMatchingController(t.Context(), c, "io.cilium/gateway-controller", logger)
-
-	t.Run("invalid object", func(t *testing.T) {
-		res := fn(&corev1.Pod{})
-		require.False(t, res)
-	})
-
-	t.Run("gateway is matched by controller", func(t *testing.T) {
-		res := fn(&gatewayv1.Gateway{
-			Spec: gatewayv1.GatewaySpec{
-				GatewayClassName: "cilium",
-			},
-		})
-		require.True(t, res)
-	})
-
-	t.Run("gateway is linked to non-existent class", func(t *testing.T) {
-		res := fn(&gatewayv1.Gateway{
-			Spec: gatewayv1.GatewaySpec{
-				GatewayClassName: "non-existent",
-			},
-		})
-		require.False(t, res)
-	})
-}
-
-func Test_gatewayReconcilePredicate(t *testing.T) {
-	logger := hivetest.Logger(t)
-	c := fake.NewClientBuilder().WithScheme(helpers.TestScheme(helpers.AllOptionalKinds)).WithObjects(controllerTestFixture...).Build()
-	predicate := gatewayOwnedByController(hasMatchingController(t.Context(), c, controllerName, logger))
-
-	t.Run("update keeps handoff reconcile when old gateway matched", func(t *testing.T) {
-		require.True(t, predicate.Update(event.UpdateEvent{
-			ObjectOld: &gatewayv1.Gateway{
-				Spec: gatewayv1.GatewaySpec{GatewayClassName: "cilium"},
-			},
-			ObjectNew: &gatewayv1.Gateway{
-				Spec: gatewayv1.GatewaySpec{GatewayClassName: "non-existent"},
-			},
-		}))
-	})
-
-	t.Run("create ignores gateway that never matched", func(t *testing.T) {
-		require.False(t, predicate.Create(event.CreateEvent{
-			Object: &gatewayv1.Gateway{
-				Spec: gatewayv1.GatewaySpec{GatewayClassName: "non-existent"},
-			},
-		}))
-	})
-}
-
-func Test_gatewayClassReconcilePredicate(t *testing.T) {
-	predicate := gatewayClassOwnedByController(controllerName)
-
-	t.Run("update keeps handoff reconcile when old gatewayclass matched", func(t *testing.T) {
-		require.True(t, predicate.Update(event.UpdateEvent{
-			ObjectOld: &gatewayv1.GatewayClass{
-				Spec: gatewayv1.GatewayClassSpec{ControllerName: controllerName},
-			},
-			ObjectNew: &gatewayv1.GatewayClass{
-				Spec: gatewayv1.GatewayClassSpec{ControllerName: "example.com/other-controller"},
-			},
-		}))
-	})
-
-	t.Run("create ignores unrelated gatewayclass", func(t *testing.T) {
-		require.False(t, predicate.Create(event.CreateEvent{
-			Object: &gatewayv1.GatewayClass{
-				Spec: gatewayv1.GatewayClassSpec{ControllerName: "example.com/other-controller"},
-			},
-		}))
-	})
-}
 
 func Test_success(t *testing.T) {
 	tests := []struct {
