+++ operator/pkg/gateway-api/gateway.go
@@ -14,13 +14,13 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 	mcsapiv1beta1 "sigs.k8s.io/mcs-api/pkg/apis/v1beta1"
 
 	"github.com/cilium/cilium/operator/pkg/gateway-api/helpers"
 	"github.com/cilium/cilium/operator/pkg/gateway-api/indexers"
+	"github.com/cilium/cilium/operator/pkg/gateway-api/predicates"
 	watchhandlers "github.com/cilium/cilium/operator/pkg/gateway-api/watch-handlers"
 	"github.com/cilium/cilium/operator/pkg/model/translation"
 	ciliumv2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
@@ -89,7 +89,7 @@ func (r *gatewayReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	}
 
 	// Index Gateways by implementation (ie `cilium`)
-	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &gatewayv1.Gateway{}, indexers.ImplementationGatewayIndex, indexers.GenerateIndexerGatewayByImplementation(r.Client, controllerName)); err != nil {
+	if err := mgr.GetFieldIndexer().IndexField(context.Background(), &gatewayv1.Gateway{}, indexers.ImplementationGatewayIndex, indexers.GenerateIndexerGatewayByImplementation(r.Client, helpers.CiliumDefaultControllerName)); err != nil {
 		return fmt.Errorf("failed to setup field indexer %q: %w", indexers.ImplementationGatewayIndex, err)
 	}
 
@@ -120,26 +120,26 @@ func (r *gatewayReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		return fmt.Errorf("failed to setup field indexer %q: %w", indexers.BackendTLSPolicyConfigMapIndex, err)
 	}
 
-	hasMatchingControllerFn := hasMatchingController(context.Background(), r.Client, controllerName, r.logger)
+	hasMatchingControllerFn := helpers.GatewayHasMatchingControllerFn(context.Background(), r.Client, helpers.CiliumDefaultControllerName, r.logger)
 	gatewayBuilder := ctrl.NewControllerManagedBy(mgr).
 		// Watch its own resource
 		For(&gatewayv1.Gateway{},
-			builder.WithPredicates(gatewayOwnedByController(hasMatchingControllerFn))).
+			builder.WithPredicates(predicates.GatewayOwnedByController(hasMatchingControllerFn))).
 		// Watch GatewayClass resources, which are linked to Gateway
 		Watches(&gatewayv1.GatewayClass{},
 			watchhandlers.EnqueueRequestForOwningGatewayClass(r.Client, *r.logger),
-			builder.WithPredicates(gatewayClassOwnedByController(controllerName))).
+			builder.WithPredicates(predicates.GatewayClassOwnedByController(helpers.CiliumDefaultControllerName))).
 		// Watch related backend Service for status
 		// LB Services are handled by the Owns call later.
 		Watches(&corev1.Service{}, watchhandlers.EnqueueRequestForBackendService(r.Client, *r.logger)).
 		// Watch HTTPRoute linked to Gateway
-		Watches(&gatewayv1.HTTPRoute{}, watchhandlers.EnqueueRequestForOwningHTTPRoute(r.Client, r.logger, controllerName)).
+		Watches(&gatewayv1.HTTPRoute{}, watchhandlers.EnqueueRequestForOwningHTTPRoute(r.Client, r.logger, helpers.CiliumDefaultControllerName)).
 		// Watch GRPCRoute linked to Gateway
-		Watches(&gatewayv1.GRPCRoute{}, watchhandlers.EnqueueRequestForOwningGRPCRoute(r.Client, r.logger, controllerName)).
+		Watches(&gatewayv1.GRPCRoute{}, watchhandlers.EnqueueRequestForOwningGRPCRoute(r.Client, r.logger, helpers.CiliumDefaultControllerName)).
 		// Watch related secrets used to configure TLS
 		Watches(&corev1.Secret{},
 			watchhandlers.EnqueueRequestForTLSSecret(r.Client, r.logger),
-			builder.WithPredicates(predicate.NewPredicateFuncs(r.usedInGateway))).
+			builder.WithPredicates(predicate.NewPredicateFuncs(predicates.SecretUsedInGatewayFn(r.Client, r.logger)))).
 		// Watch related namespace in allowed namespaces
 		Watches(&corev1.Namespace{},
 			watchhandlers.EnqueueRequestForAllowedNamespace(r.Client, r.logger)).
@@ -154,7 +154,7 @@ func (r *gatewayReconciler) SetupWithManager(mgr ctrl.Manager) error {
 
 	if tlsRouteEnabled {
 		// Watch TLSRoute linked to Gateway
-		gatewayBuilder = gatewayBuilder.Watches(&gatewayv1.TLSRoute{}, watchhandlers.EnqueueRequestForOwningTLSRoute(r.Client, r.logger, controllerName))
+		gatewayBuilder = gatewayBuilder.Watches(&gatewayv1.TLSRoute{}, watchhandlers.EnqueueRequestForOwningTLSRoute(r.Client, r.logger, helpers.CiliumDefaultControllerName))
 	}
 
 	if serviceImportEnabled {
@@ -163,47 +163,4 @@ func (r *gatewayReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	}
 
 	return gatewayBuilder.Complete(r)
-}
-
-func gatewayOwnedByController(hasMatchingControllerFn func(object client.Object) bool) predicate.Predicate {
-	return predicate.Funcs{
-		CreateFunc: func(e event.CreateEvent) bool {
-			return hasMatchingControllerFn(e.Object)
-		},
-		DeleteFunc: func(e event.DeleteEvent) bool {
-			return hasMatchingControllerFn(e.Object)
-		},
-		UpdateFunc: func(e event.UpdateEvent) bool {
-			// Reconcile one last time when a Gateway moves away from this controller
-			// so previously managed resources can be cleaned up.
-			return hasMatchingControllerFn(e.ObjectOld) || hasMatchingControllerFn(e.ObjectNew)
-		},
-		GenericFunc: func(e event.GenericEvent) bool {
-			return hasMatchingControllerFn(e.Object)
-		},
-	}
-}
-
-func gatewayClassOwnedByController(controllerName string) predicate.Predicate {
-	hasMatchingControllerName := matchesControllerName(controllerName)
-	return predicate.Funcs{
-		CreateFunc: func(e event.CreateEvent) bool {
-			return hasMatchingControllerName(e.Object)
-		},
-		DeleteFunc: func(e event.DeleteEvent) bool {
-			return hasMatchingControllerName(e.Object)
-		},
-		UpdateFunc: func(e event.UpdateEvent) bool {
-			// Reconcile one last time when a GatewayClass moves away from this controller
-			// so referenced Gateways can clean up previously managed resources.
-			return hasMatchingControllerName(e.ObjectOld) || hasMatchingControllerName(e.ObjectNew)
-		},
-		GenericFunc: func(e event.GenericEvent) bool {
-			return hasMatchingControllerName(e.Object)
-		},
-	}
-}
-
-func (r *gatewayReconciler) usedInGateway(obj client.Object) bool {
-	return len(helpers.GetGatewaysForSecret(context.Background(), r.Client, obj, r.logger)) > 0
 }
