+++ operator/pkg/gateway-api/helpers/consts.go
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package helpers
+
+const (
+	// CiliumDefaultControllerName is the gateway controller name used in cilium.
+	CiliumDefaultControllerName = "io.cilium/gateway-controller"
+)
