+++ operator/pkg/gateway-api/helpers/gateway.go
@@ -1,0 +1,36 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package helpers
+
+import (
+	"context"
+	"log/slog"
+
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
+
+	"github.com/cilium/cilium/pkg/logging/logfields"
+)
+
+func GatewayHasMatchingControllerFn(ctx context.Context, c client.Client, controllerName string, logger *slog.Logger) func(object client.Object) bool {
+	return func(obj client.Object) bool {
+		scopedLog := logger.With(
+			logfields.Resource, obj.GetName(),
+		)
+		gw, ok := obj.(*gatewayv1.Gateway)
+		if !ok {
+			return false
+		}
+
+		gwc := &gatewayv1.GatewayClass{}
+		key := types.NamespacedName{Name: string(gw.Spec.GatewayClassName)}
+		if err := c.Get(ctx, key, gwc); err != nil {
+			scopedLog.ErrorContext(ctx, "Unable to get GatewayClass", logfields.Error, err)
+			return false
+		}
+
+		return string(gwc.Spec.ControllerName) == controllerName
+	}
+}
